+++ bible/src/main/java/com/library/bible/member/service/IMemberService.java
@@ -5,10 +5,10 @@ import java.util.List;
 import com.library.bible.member.model.Member;
 
 public interface IMemberService {
-    Member selectMember(int memId);
+    Member selectMember(long memId);
     Member selectMemberByMemEmail(String memEmail);
     List<Member> selectAllMembers();
     Member insertMember(Member member, String role);
     Member updateMember(Member member);
-    void deleteMember(int memId);
+    void deleteMember(long memId);
 }
