+++ bible/src/main/java/com/library/bible/memberrent/service/IMemberRentService.java
@@ -3,8 +3,8 @@ package com.library.bible.memberrent.service;
 import com.library.bible.memberrent.model.MemberRent;
 
 public interface IMemberRentService {
-    MemberRent selectMemberRentByMemId(int memId);
+    MemberRent selectMemberRentByMemId(long memId);
     int insertMemberRent(MemberRent memberRent);
     int updateMemberRent(MemberRent memberRent);
-    void deleteMemberRent(int memId);
+    void deleteMemberRent(long memId);
 }
